Implement multiple tags filtering per wallet endpoints
