+++ app/src/Controller/Wallets/IndexController.php
@@ -4,15 +4,11 @@ declare(strict_types=1);
 
 namespace App\Controller\Wallets;
 
-use App\Database\Charge;
 use App\Database\Wallet;
-use App\Repository\ChargeRepository;
 use App\Repository\WalletRepository;
-use App\Service\ChargeWalletService;
 use App\View\WalletView;
 use Psr\Http\Message\ResponseInterface;
 use Spiral\Auth\AuthScope;
-use Spiral\Http\Request\InputManager;
 use Spiral\Http\ResponseWrapper;
 use Spiral\Router\Annotation\Route;
 
@@ -23,8 +19,6 @@ final class IndexController extends Controller
         private ResponseWrapper $response,
         private WalletRepository $walletRepository,
         private WalletView $walletView,
-        private ChargeRepository $chargeRepository,
-        private ChargeWalletService $chargeWalletService,
     ) {
         parent::__construct($auth);
     }
@@ -39,28 +33,5 @@ final class IndexController extends Controller
         }
 
         return $this->walletView->json($wallet);
-    }
-
-    #[Route(route: '/wallets/<id>/total', name: 'wallet.index.total', methods: 'GET', group: 'auth')]
-    public function indexTotal(string $id, InputManager $input): ResponseInterface
-    {
-        $wallet = $this->walletRepository->findByPKByUserPK((int) $id, (int) $this->user->id);
-
-        if (! $wallet instanceof Wallet) {
-            return $this->response->create(404);
-        }
-
-        $this->chargeRepository->filter($input->query->fetch(['date-from', 'date-to']));
-
-        $income = $this->chargeRepository->totalByWalletPK((int) $wallet->id, Charge::TYPE_INCOME);
-        $expense = $this->chargeRepository->totalByWalletPK((int) $wallet->id, Charge::TYPE_EXPENSE);
-
-        return $this->response->json([
-            'data' => [
-                'totalAmount' => $this->chargeWalletService->totalByIncomeAndExpense($income, $expense),
-                'totalIncomeAmount' => $income,
-                'totalExpenseAmount' => $expense,
-            ],
-        ]);
     }
 }
