+++ app/src/Service/Statistics/ChargeAmountGraph.php
@@ -11,6 +11,7 @@ use App\Repository\ChargeRepository;
 use App\Service\Filter\Filter;
 use App\Service\Filter\FilterType;
 use Cycle\Database\Injection\Fragment;
+use Cycle\Database\Injection\Parameter;
 use Cycle\Database\Query\SelectQuery;
 
 class ChargeAmountGraph
@@ -21,6 +22,8 @@ class ChargeAmountGraph
     use Filter;
 
     protected Group $grouping = Group::ByMonth;
+
+    protected array $tagIds = [];
 
     public function __construct(
         private readonly ChargeRepository $chargeRepository
@@ -34,12 +37,20 @@ class ChargeAmountGraph
         return $this;
     }
 
+    public function groupByTags(array $ids): static
+    {
+        $this->tagIds = $ids;
+
+        return $this;
+    }
+
     public function getGraph(Tag $tag = null, Wallet $wallet = null): array
     {
         $query = $this->buildQueryByTagAndWallet($tag, $wallet);
 
         $data = new ChargeAmountData($this->grouping);
         $data->filter($this->filter);
+        $data->setTagIds($this->tagIds);
         $data->setIncome((clone $query)->where('charges.type', Charge::TYPE_INCOME)->fetchAll());
         $data->setExpense((clone $query)->where('charges.type', Charge::TYPE_EXPENSE)->fetchAll());
 
@@ -52,6 +63,8 @@ class ChargeAmountGraph
 
         if ($tag !== null) {
             $query = $query->where('tag_charges.tag_id', $tag->id);
+        } else if (count($this->tagIds) > 0) {
+            $query = $query->where('tag_charges.tag_id', 'in', new Parameter($this->tagIds));
         }
 
         if ($wallet !== null) {
@@ -65,14 +78,24 @@ class ChargeAmountGraph
     {
         $query = $this->chargeRepository->select()->buildQuery();
 
+        $columns = [
+            $this->grouping->getQueryFragment('charges.created_at'),
+            new Fragment('SUM(charges.amount) AS total'),
+        ];
+
+        if (count($this->tagIds) > 0) {
+            $columns[] = new Fragment('tag_charges.tag_id AS tag_id');
+        }
+
         $query = $query->from('charges')
-                       ->columns([
-                           $this->grouping->getQueryFragment('charges.created_at'),
-                           new Fragment('SUM(charges.amount) AS total')
-                       ])
+                       ->columns($columns)
                        ->leftJoin('tag_charges')
                        ->on('tag_charges.charge_id', 'charges.id')
                        ->groupBy('date');
+
+        if (count($this->tagIds) > 0) {
+            $query = $query->groupBy('tag_id');
+        }
 
         $this->injectFilter($query);
