+++ tests/Feature/Controller/Wallets/TagsControllerTest.php
@@ -105,221 +105,6 @@ class TagsControllerTest extends TestCase implements DatabaseTransaction
         $this->assertArrayNotContains($tag3->name, $body, 'data.*.name');
     }
 
-    public function testTotalRequireAuth(): void
-    {
-        $user = $this->userFactory->create();
-        $wallet = $this->walletFactory->forUser($user)->create();
-        $tag = $this->tagFactory->forUser($user)->create();
-        $this->chargeFactory->forUser($user)->forWallet($wallet)->withTags([$tag])->create();
-
-        $response = $this->get("/wallets/{$wallet->id}/tags/{$tag->id}/total");
-
-        $response->assertUnauthorized();
-    }
-
-    public function testTotalOfMissingTagNotFound(): void
-    {
-        $auth = $this->makeAuth($user = $this->userFactory->create());
-        $wallet = $this->walletFactory->forUser($user)->create();
-        $tagId = Fixtures::integer();
-
-        $response = $this->withAuth($auth)->get("/wallets/{$wallet->id}/tags/{$tagId}/total");
-
-        $response->assertNotFound();
-    }
-
-    public function testTotalOfMissingWalletAndTagNotFound(): void
-    {
-        $auth = $this->makeAuth($this->userFactory->create());
-        $walletId = Fixtures::integer();
-        $tagId = Fixtures::integer();
-
-        $response = $this->withAuth($auth)->get("/wallets/{$walletId}/tags/{$tagId}/total");
-
-        $response->assertNotFound();
-    }
-
-    public function testTotalOfForeignWalletReturnNotFound(): void
-    {
-        $auth = $this->makeAuth($this->userFactory->create());
-        $wallet = $this->walletFactory->forUser($otherUser = $this->userFactory->create())->create();
-        $tag = $this->tagFactory->forUser($otherUser)->create();
-        $this->chargeFactory->forUser($otherUser)->forWallet($wallet)->withTags([$tag])->create();
-
-        $response = $this->withAuth($auth)->get("/wallets/{$wallet->id}/tags/{$tag->id}/total");
-
-        $response->assertNotFound();
-    }
-
-    public function testTotalOfForeignTagReturnNotFound(): void
-    {
-        $auth = $this->makeAuth($user = $this->userFactory->create());
-        $wallet = $this->walletFactory->forUser($user)->create();
-        $tag = $this->tagFactory->forUser($this->userFactory->create())->create();
-        $this->chargeFactory->forUser($user)->forWallet($wallet)->withTags([$tag])->create();
-
-        $response = $this->withAuth($auth)->get("/wallets/{$wallet->id}/tags/{$tag->id}/total");
-
-        $response->assertNotFound();
-    }
-
-    public function testTotalOfNoChargesWithTag(): void
-    {
-        $auth = $this->makeAuth($user = $this->userFactory->create());
-        $wallet = $this->walletFactory->forUser($user)->create();
-        $this->chargeFactory->forUser($user)->forWallet($wallet)->createMany(3);
-        $tag = $this->tagFactory->forUser($user)->create();
-
-        $response = $this->withAuth($auth)->get("/wallets/{$wallet->id}/tags/{$tag->id}/total");
-
-        $response->assertOk();
-
-        $body = $this->getJsonResponseBody($response);
-
-        $this->assertArrayContains(0, $body, 'data.totalAmount');
-        $this->assertArrayContains(0, $body, 'data.totalIncomeAmount');
-        $this->assertArrayContains(0, $body, 'data.totalExpenseAmount');
-    }
-
-    public function testTotalReturnsTotalByTag(): void
-    {
-        $auth = $this->makeAuth($user = $this->userFactory->create());
-        $wallet = $this->walletFactory->forUser($user)->create();
-        $tag = $this->tagFactory->forUser($user)->create();
-
-        $this->chargeFactory->forUser($user)->forWallet($wallet)->createMany(4);
-
-        $chargesAmount = rand(5, 20);
-        $totalIncome = 0.00;
-        $totalExpense = 0.00;
-
-        $this->chargeFactory->withTags([$tag]);
-
-        for ($i = 0; $i < $chargesAmount; $i++) {
-            $charge = $this->chargeFactory->create();
-
-            if ($charge->type === Charge::TYPE_INCOME) {
-                $totalIncome += $charge->amount;
-            } else {
-                $totalExpense += $charge->amount;
-            }
-        }
-
-        $total = round($totalIncome - $totalExpense, 2);
-
-        $response = $this->withAuth($auth)->get("/wallets/{$wallet->id}/tags/{$tag->id}/total");
-
-        $response->assertOk();
-
-        $body = $this->getJsonResponseBody($response);
-
-        $this->assertArrayContains($total, $body, 'data.totalAmount');
-        $this->assertArrayContains($totalIncome, $body, 'data.totalIncomeAmount');
-        $this->assertArrayContains($totalExpense, $body, 'data.totalExpenseAmount');
-    }
-
-    public function totalWithDateFiltersReturnsFilteredTotalByTagDataProvider(): array
-    {
-        return [
-            [
-                [
-                    'total' => 200,
-                    'income' => 410,
-                    'expense' => 210,
-                ],
-                [
-                    'date-from' => '00-13-2022',
-                    'date-to' => '40-00-2022',
-                ]
-            ],
-            [
-                [
-                    'total' => 200,
-                    'income' => 410,
-                    'expense' => 210,
-                ],
-                [
-                    'date-from' => '01-06-2022',
-                    'date-to' => '04-06-2022',
-                ]
-            ],
-            [
-                [
-                    'total' => 100,
-                    'income' => 205,
-                    'expense' => 105,
-                ],
-                [
-                    'date-from' => '02-06-2022',
-                    'date-to' => '03-06-2022',
-                ]
-            ],
-            [
-                [
-                    'total' => 150,
-                    'income' => 309,
-                    'expense' => 159,
-                ],
-                [
-                    'date-from' => '02-06-2022',
-                ]
-            ],
-            [
-                [
-                    'total' => 150,
-                    'income' => 306,
-                    'expense' => 156,
-                ],
-                [
-                    'date-to' => '03-06-2022',
-                ]
-            ],
-        ];
-    }
-
-    /**
-     * @dataProvider totalWithDateFiltersReturnsFilteredTotalByTagDataProvider
-     * @param array $total
-     * @param array $query
-     * @return void
-     */
-    public function testTotalWithDateFiltersReturnsFilteredTotalByTag(array $total, array $query): void
-    {
-        $auth = $this->makeAuth($user = $this->userFactory->create());
-        $wallet = $this->walletFactory->forUser($user)->create();
-        $tag = $this->tagFactory->forUser($user)->create();
-
-        for ($i = 1; $i <= 4; $i++) {
-            $charge = ChargeFactory::make();
-            $charge->type = Charge::TYPE_INCOME;
-            $charge->amount = 100 + $i;
-            $charge->createdAt = new \DateTimeImmutable("0{$i}-06-2022");
-            $this->chargeFactory->forUser($user)->forWallet($wallet)->withTags([$tag])->create($charge);
-        }
-
-        for ($i = 1; $i <= 4; $i++) {
-            $charge = ChargeFactory::make();
-            $charge->type = Charge::TYPE_EXPENSE;
-            $charge->amount = 50 + $i;
-            $charge->createdAt = new \DateTimeImmutable("0{$i}-06-2022");
-            $this->chargeFactory->forUser($user)->forWallet($wallet)->withTags([$tag])->create($charge);
-        }
-
-        $this->chargeFactory->forUser($user)->forWallet($wallet)->withTags([
-            $this->tagFactory->forUser($user)->create()
-        ])->createMany(4);
-
-        $response = $this->withAuth($auth)->get("/wallets/{$wallet->id}/tags/{$tag->id}/total", $query);
-
-        $response->assertOk();
-
-        $body = $this->getJsonResponseBody($response);
-
-        $this->assertArrayContains($total['total'], $body, 'data.totalAmount');
-        $this->assertArrayContains($total['income'], $body, 'data.totalIncomeAmount');
-        $this->assertArrayContains($total['expense'], $body, 'data.totalExpenseAmount');
-    }
-
     public function testFindRequireAuth(): void
     {
         $wallet = $this->walletFactory->create();
